Updates
Update my google password encryption for mail credentials.

Bring back the PR notification thingy..The double messages and all that was an error on my end due to the fact of the damn appveyor bot not actually setting up webhooks correctly after I reauthed it recently.. it is now reset up again, with new project files on appveyor.