Updates
Main thing is moving the folder destination for tinyMediaManager, due to not being able to be run from %ChocolateyInstall% with the error "Cannot write to TMM directory, have no rights - exiting." ... moving this to "C:\tools\tmm" allows it to run.

Dunno how I never caught this error till now but I have to thank "ksz16" on chocolatey.org as he is the one who brought it to my attention.